refactor: use uint for array indices and update property removal API
Changes the type of array index parameters from int to uint in jsonpath functions to ensure indices are non-negative. Renames RemovePropertiesByName to RemovePropertiesByPath and updates the method to accept Path objects instead of strings, improving type safety and API consistency. Converts NewPath from a variable assignment to a proper function wrapper.

+++ pkg/govy/rules_for_slice.go
@@ -164,7 +164,7 @@ func (r PropertyRulesForSlice[S, T, P]) plan(builder planBuilder) {
 
 // getJSONPathForIndex returns a JSONPath for the given index.
 func (r PropertyRulesForSlice[S, T, P]) getJSONPathForIndex(index int) string {
-	return jsonpath.JoinArray(r.sliceRules.getName(), jsonpath.NewArrayIndex(index))
+	return jsonpath.JoinArray(r.sliceRules.getName(), jsonpath.NewArrayIndex(uint(index))) // #nosec G115
 }
 
 // getName returns the name of the property.
